multiple withdraw test
